Refactor chat history table layout
