+++ views/pdf.handlebars
@@ -4,7 +4,7 @@
       style="display:flex; flex-direction: column; justify-content: center; align-items: center;"
     >
       <h2>Vestlusajalugu</h2>
-      <table id="chatHistoryTable"></table>
+      <table id="chatHistoryTable" style="padding: 16px;"></table>
     </div>
   </body>
 </html>
